Core, Dom, DragDrop, Validation
Core: New functions (Extends and InstanceOf) allows for developers to
check whether a given object extends a given class, or is an instance
or extension of a given class. Also improved CreateOverride which now
allows for override function to return a value, and made it work
properly with IE8.

Dom: Cleanup

DragDrop: Fixed broken event handling

Validation: ExpectInstance(..) now accepts object if it is either an
instance of the given type, or an extension of the given type.