DropDown, PickerBase, TreeView
Fixed problem with DropDown in Chrome which scrolled nodes to the top
if selected using a checkbox.
The scrolling happend because the picker control (TreeView) temporarily
lost “picker mode” due to lose of focus when checkbox was clicked. Once
fixed another bug was revealed which was introduced in commit
https://github.com/Jemt/Fit.UI/commit/50154e1be8a14cc4e5ebe39994ae85a092
eeee8e causing incorrect scrolling when using arrow keys in the
DropDown control.
Some cleanup was also performed, including getting rid of unused code,
and getting rid of the terrible SetEventDispatcher function of
PickerBase which is no longer used.