DropDown
DropDown no longer closes on Desktop when focus is lost. Behaviour
changed as it didn’t work well when a ContextMenu was opened inside the
DropDown as it would close, leaving the ContextMenu floating without
any visual context. It also made debugging a lot harder as it prevented
the DropDown from staying open when interacting with Developer Tools.