AddHandler: Now iFrame OnLoad only fires immediately if "src" is set.
ReadyState property for contentDocument on dynamically created iFrames return "complete" when a src attribute is not set, but often src is set "later" when needed, making the old behaviour (where onload is fired immediately if src attribute is not set) erroneous.
See code comments for more information.
This fixes #26.

Also fixed bug in IE where accessing frame.contentDocument resulted in "unspecified error" (IE8) or Null (IE9) if iFrame was not rooted in DOM.