TreeView as picker: now resetting focus when reopening DropDown.
Using TreeView in DropDown caused a node previously selected using the keyboard to remain focused after closing drop down. This was annoying as reopening the drop down using the keyboard caused navigation to start where the last node was selected, even though the drop down was scrolled to the top.