#54 Fixed inconsistency in how TreeViewNode titles were handled.
WSTreeView with an empty preselection ( e.g. ctl.SetSelections([{Title: "", Value: ""}]); ) could cause a type check error to be thrown once data was loaded, since the preselection's empty title would be passed to TreeView.Title(..) which in turn would cause an exception to be thrown.

In general we should avoid using ExpectStringValue, but use ExpectString instead. ExpectStringValue goes beyond just type checking which is not desirable.