Clarified what DOMElement means in Fit.UI.
When we expect a DOMElement, we actually expect an instance of the HTMLElement type, not the Element type. But due to support for legacy IE, we previously checked against the Element interface since legacy IE does not have the HTMLElement interface. This caused some confusion, and the posibility of passing XML elements which is not something Fit.UI should be able to handle.