#77, #78, #79, #80, and various improvements.
#77: Fixed bug in Selectable() which removed preselections from WSTreeView.

#78: Fixed bug in RemoveAllChildren(true) which did not dispose node objects, but only removed them from TreeView. Fixing this broke Reload function which relied on a certain approach to preserve selections when reloading (var sel = tv.Selected();tv.Reload();tv.Selected(sel);). This was no longer possible with disposed nodes, so a new implementation of Reload has been made which addresses this problem.
This is a breaking change! Now use tv.Reload(true) to preserve selections.

#79: Fixed bugs related to the potential problem of nodes having been disposed or detached while loading data from WebService.

#80: Fixed Selected(newSelection) which did not fire OnChange.

Also improved EnsureData which only ensured data if not previously loaded. But we might want to use it to re-ensure all nodes after calling RemoveAllChildren(..) which is now possible. Data will be ensured if no children are contained in TreeView.

Various other minor improvements.