DropDown: Fixed bug causing DropDownMaxWidth to only work once.
DropDown picker is supposed to adjust to the width of selectable items, up to the maximum width set using DropDownMaxWidth(..). However, the drop down menu only assumed the width of the content the first time it was opened, reverting to the width of the DropDown control itself when closed.
The bug was introduced with the DetectBoundaries feature.