#57: DropDown now opens when clicked in TextSelectionMode + other fixes.
Clicking the selection area in a DropDown in TextSelectionMode now results in DropDown being opened (#57). Previously only the arrow button did this.

The fix to #57 also results in focus being preserved when toggling items in picker control on mobile, when DropDown is initially focused by clicking an already selected item. Previously it only preserved focus when user clicked the input area (itemContainer) of the DropDown control.

Also fixed optimizeDropDownPosition which failed when DetectBoundaries was enabled and items were being seleted programmatically while control was hidden, which is a common approach to reduce reflows when selecting large amounts of items.

Furthermore made minor change to onclick event for arrow button which suppressed event propagation when DropDown was opened, but not when closed. Made it always suppress event bubbling so the event never reaches itemContainer.onclick, which allowed for additional simplification and ensured consistency.