WSDropDown: Search value no longer passed to OnRequest handler on reload
Fixed bug resulting in search value being passed to OnRequest handler when reloading data for TreeView. Search value is only intended for ListView and should not affect TreeView data which is expected to expose the entire data set - although not necessarily return it all at once.

Also fixed bug resulting in ClearData(..) and AutoUpdateSelected(..) not waiting for ListView data to finish loading.