Input: Fixed DesignMode when not rooted and editor already loaded.
For an Input control not rooted in DOM, and CKEditor already loaded on the page, resulted in DesignMode never being enabled when control was rooted in DOM. Bug could easily occur in a React application where multiple controls are initialized over time, and only rendered when needed. Bug was caused by DesignMode() using designEditor variable to determine whether DesignMode was enabled or not - but this variable is not reliable as it may be set async.

Also made sure to update DesignMode state earlier when disabling it, to make sure MultiLine(..) behaves properly (does not try to disable DesignMode again which will result in an error due to already disposed objects).