DropDown: Fixed Focused(true) for control with partially hidden value.
DropDown was incorrectly rendered if Focused(true) was called for a DropDown in single selection mode with an item partially hidden. The primary (right) input control would always be focused, even when an item was selected and partially hidden (along with the primary input field), which would cause the hidden input field to be scrolled into view, breaking the visual representation of the control.