Typings generator: Added support for generics in callbacks.
Generics previously only worked if used as a return type. Now we can declare a parameter that takes a generic type that is passed to an anonymous function (callback) defined in another parameter.

This commit also fixes support for anonymous functions (callbacks) which would previously not have had their parameter types resolved to typescript equivalents.