Typings generator: Improved support for generics and dictionaries.
Generics can now be passed to classes, allowing function X to return an instance of a class with properties using types passed to function X.

Dictionaries and generics are now described in the proper JSDoc format.

Fixed bug which caused a single generic type not to transpile when used in an inline callback.

Fixed bug related to the use of multiple possible types of which one or more was contained in a namespace, due to the use of the period character which was not previously allowed:
string | Fit.Demo.MyType | null