Typings generator: Added support for genric type $TypeOfThis.
The special generic type $TypeOfThis allows us to pass the type of 'this' to callbacks and DTOs. This is required for inherited functions such as ControlBase.OnChange to provide the specialized type of a control to the event handler, rather than the type from which the function is inherited (e.g. ControlBase).