#117: Fixed bugs related to initialization while hidden.
Programmatically focusing DropDown immediately after making it visible no longer results in a render bug where the wrong input field is "scrolled into view" if a partially hidden item exists.

If input value is set while control is hidden, and a partially hidden item exists, input value is now assigned to the DropDown's left side input control instead of the right side input control, which won't be visible.
Also fixed bug which caused the input value to "stick" in this case - it's supposed to be cleared when the control receives focus for the first time.