#120: DropDown - selected items no longer becomes partially hidden.
This resolves the problem with delete icon being pushed "out of view" due to overflow, which made it difficult for the user to remove such items.

These changes also resolves #118 with the primary input field sometimes becoming too narrow to be usable.

Furthermore the logic required to adjust the width of input fields as the user types, is no longer needed, so these changes also resolves the problem mentioned in #113.