DOM: Clarified and fixed GetScrollParent and GetOverflowingParent.
Previously GetScrollParent and GetOverflowingParent had almost identical implementations, with the exception of GetScrollParent only considering overflow:auto|scroll, while GetOverflowingParent also considered overflow:hidden. GetOverflowingParent also had a partially buggy implementation due to the use of an 'elm' variable not available in a loop, which actually made the buggy version work almost like the updated version that comes with this commit.

Both functions were designed in a way that caused them to ignore a parent with overflow set, if the element passed had position:absolute and the scrolling/overflowing parent was not positioned to "anchor" the element (become offsetParent to the element) - this was the part that actually didn't work for GetOverflowingParent.
This behaviour makes sense for GetScrollParent as its intended behaviour was to find the parent element with scroll that affects the position of the element when scrolled. But the implementation was buggy as it did not consider the fact that any child within the scrolling parent could act as anchor (offsetParent) for the element passed to the function, hence make it scroll along with the content of the scroll parent.
For GetOverflowingParent this behaviour does not make sense. We want this function to return the first parent that has overflow set, no matter if it affects positioning or not when scrolled. But at the same time we want to be able to use it to identify a potential scroll parent by allowing overflow:hidden to be ignored.

With this commit GetScrollParent now truly returns the parent that affects the position of the element passed if scrolled, and GetOverflowingParent returns the first parent found up the hierarchy with overflow set, without considering anchor/offset relations.