Input: Improved lazy validation.
Controls with LazyValidation(true) now updates visual valid state when value is updated programmatically if control at some point had focus and lost it again (i.e. has been interacted with by the user).

A popular use case for this is a control removing whitepaces OnBlur which might affect valid state, so we need to make sure the UI reflects this.