Input: Editor dialogs are once again mounted locally next to control.
It is causing too many problems with the dialog being placed in the global scope - especially with light dismissable callouts and panels which close when interacting with the dialog. With this commit a dialog is yet again mounted next to the Input control. This is made possible with a tweak to the base64image plugin which no longer uses CKEditor's file picker, as this previously prevented us from moving the dialog DOM element.

Furthermore improved the image resize plugin so it no longer leaves a resize handler on the screen if HTML editor is disposed OnBlur.