#106: Fixed mechanism preventing DropDown from losing focus on scroll.
This is a fix to a fix preventing DropDown (TreeView picker actually) from losing focus to a parent with tabIndex set when scrolling TreeView picker. The original fix assumed tabIndex would have a value of 0 or greater, but it must support a value of -1 as well which is used to allow programmatically assigned focus.