WSDropDown: Several fixes to new action menu.
"Search for more options" (WSListView) is no longer added to action menu if InputEnabled is false.

"Show all available options" (WSTreeView) now reloads data if ClearData() has been invoked.

"Show all available options" (WSTreeView) immediately returned to action menu when data was loaded if UseActionMenu(true) was used (enforced). Fixed - TreeView data is now shown instead. However, if no children is returned, it still switches back to the action menu, which now indicates that no data is available in the TreeView. Prior to this change the user just got an empty picker.

"Search for more options" now displays the placeholder in Visual Selection Mode, even if items are selected.

Selecting "Search for more options" in the action menu now closes the dropdown control.
If DropDown is in Visual Selection Mode, the search field is displayed on a separate "line" (it kind of word wraps). When that input field lose focus (e.g. when clicking an item in the picker control), it immediately returns to its normal state. This results in the control's height changing which in turn results in the pulldown menu moving up. So if a user clicks an item in the pulldown menu, the input lose focus, the position of the pulldown menu changes, and nothing happens from the click in the picker, because it uses a traditional onclick event to register interactions, which doesn't fire unless the mouse button is released on the same object it was pressed down on. We avoid this by closing the control. We close it in both Text Selection Mode and Visual Selection Mode for consistency.

SetInputValue(..) could bring a previously focused input field in edit mode but assign the value to the primary input field which did not adjust to the input value, making it impossible to read the value. Fixed!

And various other minor fixes and improvements.