WSDropDown: Improvements related to action menu.
Action menu no longer showing if the only action available is "Search for options". In this case the control's placeholder will show "Search..". Clicking the control will make sure the input field is focused, and that the virtual keyboard is brought up on mobile devices.

Both ListView and TreeView can now be disabled using TreeViewEnabled(false) and ListViewEnabled(false) in case either is not desired.