Input: Fixed problems with invalid dimensions in base64image plugin.
Width/height fields could be cleared or set with invalid values when "Lock Ratio" was unchecked, causing width and height to assume values of "NaNpx" in style property. Fixed!

If user changed height with "Lock Ratio" unchecked, and then enabled "Lock Ratio", the dimensions would not update to respect the aspect ratio. Fixed!