#162 Input: HTML Editor height now updated when Value(..) is used.
If editor was configured with AutoGrow enabled, and toolbar was configured with HideWhenInactive, and control was not focused (toolbar was hidden), then editor would not update its height if value was changed programmatically using Value(..) or Clear().

Furthermore simplified code responsible for updating editor dimensions "later" if control is hidden or not rooted in DOM when dimensions are changed, in which case dimensions must be updated once editor becomes visible again or rooted in DOM again.