DropDown: Fixed "jumping page" when selecting items in Safari on iOS.
In some situations the page would "jump" when selecting items in the dropdown control on a mobile device. This was caused by focus being temporarily moved to one of the DropDown control's input fields before focus was then moved again to the DropDown control's arrow icon. This caused the device's virtual keyboard to start opening but immediately disappear again.

Notice that SearchModeOnFocus(..) was moved in code but not altered. It was incorrectly placed among private functions, but it belongs to the section with public functions.

Furthermore introduced Fit.Device.OptimizeForTouch (boolean) which is to replace mobile/touch detection using Fit.Browser which relies on user agent parsing.