WSDropDown/WSTreeView: Fixed performance when applying huge preselection
WSTreeView is used as a picker control in WSDropDown. When a huge selection is made with WSDropDown (hundreds of selected items), each node value is being added to WSTreeView one at a time. If a node with a given value is not yet loaded, a preselection representing the node is added. But prior to this fix this resulted in preselections being turned into new instances of Fit.Controls.TreeViewNode every time WSTreeView.Selected(..) was called, which turned out to be extremely expensive due to the many calls to the function, as well as the fact that every time a TreeViewNode was created, it involved creation and manipulation of DOM nodes.
The problem could easily be reproduced with a WSDropDown, and the following call: wsDropDown.Value("guid1;guid2;..."), where the string passed contained 500 GUIDs.

This commit greatly improves performance when working with large selections by first and foremost avoiding re-creating the same TreeViewNode instances over and over again, but also reduces DOM usage within TreeViewNode instances.