Device: Incorrectly assumed OptimizeForTouch on hybrid devices.
The bug was caused by a redundant else-if condition with a broken media query (lacked required parentheses) which resulted in hybrid devices (usually laptops with support for both touch and mouse) optimizing Fit.UI for touch primarily, which is not the intended behaviour - we always want to optimize for a desktop experience when a mouse is present.

The bug broke DatePicker on hybrid devices which did not load the calendar widget as expected, but instead layered a hidden native Date control on top of the Fit.UI control, which could be triggered only when clicking a hidden small calendar icon, which would then produce the native calendar control on the screen.