DatePicker: Improved support for hybrid (touch primary + mouse/pen).
DatePicker now works properly on hybrid devices where touch is considered the primary pointing method (and DatePicker is therefore optimized for touch and uses native picker controls) and mouse/pen is considered the secondary pointing method. Prior to this fix the user wasn't able to open the calendar widget on desktop browsers, unless he/she was lucky enough to hit the hidden calendar icon of the native date control.