HTML Editor: Now opens a link when clicked if editor is inactive.
This change required a bug fix to the data-toolbar attribute which was not set to "false" when toolbar was being hidden again. Hovever, fixing this resulted in editor starting to "jump" a few pixels whenever toolbar was toggled on/off when editor received or lost focus, if the bootstrapck skin was being used. Therefore some changes related to padding/spacing were required to make it work properly again.