ControlBase: Validation errors now updated OnBlur, not immediately.
Often a validation error was shown while entering a value in a control with validation enabled, until the value became valid - e.g. an e-mail address which would be invalid until the full e-mail address was completed. Now, the error state is not exposed to the user until the control lose focus. Once focused again, the error state will continue to be shown until the user starts correcting the value. Both dirty state and valid state continues to be updated immediately through the API which will always reveal the truth. This change is related to UI only.

The old behaviour with immediate user feedback can be enabled using ShowValidationErrorsOnChange(true).