#202: Fixed null pointer exception in WSTreeView.
WSTreeView broken when used as a picker in DropDown with DetectBoundaries(true, true). This causes WSTreeView to be contained in a parent with position:fixed without scroll when no data is loaded. This in turned caused the error because a picker without scroll (no data) caused WSTreeView to try to obtain the scroll parent from "outside" itself, but there is no such thing when position:fixed is applied.