DatePicker: New week selection mode, and now using same picker for touch
Native date pickers on touch devices do not support week selection (it works, but the entire week is not highlighted), and they are actually harder to use. It is also buggy on iPhone (https://bugs.webkit.org/show_bug.cgi?id=281525). So we now use the same calendar widget for both desktop and touch devices by default. This also has the benefit of allowing for week numbers and alternative localization on touch devices. Basically a consistent experience across devices.