DropDown: DetectBoundaries(true, true) can now escape CSS transform.
If a dropdown is constrained by CSS properties such as animation or transform, position:fixed will not be able to "escape" such parents, resulting in elements being positioned relative to such parents, rather than to the viewport. If a constraint is detected, the dropdown menu is now mounted in the document root instead, resolving this issue.