Build: Made bash scripts more robust - will now halt on errors.
Previous commit caused an NPM package to be released without minified JS and CSS files because the build machine had not executed "npm install" prior to building the package. The lack of node modules resulted in minification failing, but the build and release process continued, resulting in an incomplete NPM package.